Rename ScoreOverlay -> HUDOverlay, move to osu.Game, make it not overridable by rulesets.
