+++ osu.Desktop.VisualTests/Tests/TestCaseScoreCounter.cs
@@ -40,7 +40,7 @@ namespace osu.Desktop.VisualTests.Tests
             };
             Add(score);
 
-            ComboCounter standardCombo = new OsuComboCounter
+            BaseComboCounter comboCounter = new ComboCounter
             {
                 Origin = Anchor.BottomLeft,
                 Anchor = Anchor.BottomLeft,
@@ -48,15 +48,15 @@ namespace osu.Desktop.VisualTests.Tests
                 Count = 0,
                 TextSize = 40,
             };
-            Add(standardCombo);
+            Add(comboCounter);
 
-            PercentageCounter accuracyCombo = new PercentageCounter
+            PercentageCounter accuracyCounter = new PercentageCounter
             {
                 Origin = Anchor.TopRight,
                 Anchor = Anchor.TopRight,
                 Position = new Vector2(-20, 60),
             };
-            Add(accuracyCombo);
+            Add(accuracyCounter);
 
             StarCounter stars = new StarCounter
             {
@@ -79,26 +79,26 @@ namespace osu.Desktop.VisualTests.Tests
             AddButton(@"Reset all", delegate
             {
                 score.Count = 0;
-                standardCombo.Count = 0;
+                comboCounter.Count = 0;
                 numerator = denominator = 0;
-                accuracyCombo.SetFraction(0, 0);
+                accuracyCounter.SetFraction(0, 0);
                 stars.Count = 0;
                 starsLabel.Text = stars.Count.ToString("0.00");
             });
 
             AddButton(@"Hit! :D", delegate
             {
-                score.Count += 300 + (ulong)(300.0 * (standardCombo.Count > 0 ? standardCombo.Count - 1 : 0) / 25.0);
-                standardCombo.Count++;
+                score.Count += 300 + (ulong)(300.0 * (comboCounter.Count > 0 ? comboCounter.Count - 1 : 0) / 25.0);
+                comboCounter.Count++;
                 numerator++; denominator++;
-                accuracyCombo.SetFraction(numerator, denominator);
+                accuracyCounter.SetFraction(numerator, denominator);
             });
 
             AddButton(@"miss...", delegate
             {
-                standardCombo.Roll();
+                comboCounter.Roll();
                 denominator++;
-                accuracyCombo.SetFraction(numerator, denominator);
+                accuracyCounter.SetFraction(numerator, denominator);
             });
 
             AddButton(@"Alter stars", delegate
@@ -110,8 +110,8 @@ namespace osu.Desktop.VisualTests.Tests
             AddButton(@"Stop counters", delegate
             {
                 score.StopRolling();
-                standardCombo.StopRolling();
-                accuracyCombo.StopRolling();
+                comboCounter.StopRolling();
+                accuracyCounter.StopRolling();
                 stars.StopAnimation();
             });
         }
