+++ osu.Game.Modes.Catch/CatchRuleset.cs
@@ -1,20 +1,19 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using System.Collections.Generic;
+using OpenTK.Input;
+using osu.Game.Beatmaps;
 using osu.Game.Graphics;
 using osu.Game.Modes.Catch.UI;
 using osu.Game.Modes.Objects;
-using osu.Game.Modes.Osu.UI;
 using osu.Game.Modes.UI;
-using osu.Game.Beatmaps;
+using osu.Game.Screens.Play;
+using System.Collections.Generic;
 
 namespace osu.Game.Modes.Catch
 {
     public class CatchRuleset : Ruleset
     {
-        public override ScoreOverlay CreateScoreOverlay() => new OsuScoreOverlay();
-
         public override HitRenderer CreateHitRendererWith(Beatmap beatmap) => new CatchHitRenderer
         {
             Beatmap = beatmap,
@@ -80,6 +79,13 @@ namespace osu.Game.Modes.Catch
         protected override PlayMode PlayMode => PlayMode.Catch;
 
         public override FontAwesome Icon => FontAwesome.fa_osu_fruits_o;
+
+        public override KeyCounter[] GameplayKeys => new KeyCounter[]
+        {
+            new KeyCounterKeyboard(Key.ShiftLeft),
+            new KeyCounterMouse(MouseButton.Left),
+            new KeyCounterMouse(MouseButton.Right)
+        };
 
         public override ScoreProcessor CreateScoreProcessor(int hitObjectCount = 0) => null;
