+++ osu.Game.Modes.Osu/osu.Game.Modes.Osu.csproj
@@ -74,7 +74,6 @@
     <Compile Include="OsuDifficultyCalculator.cs" />
     <Compile Include="OsuScore.cs" />
     <Compile Include="OsuScoreProcessor.cs" />
-    <Compile Include="UI\OsuComboCounter.cs" />
     <Compile Include="UI\OsuHitRenderer.cs" />
     <Compile Include="UI\OsuPlayfield.cs" />
     <Compile Include="OsuRuleset.cs" />
@@ -85,7 +84,6 @@
     <Compile Include="Objects\Slider.cs" />
     <Compile Include="Objects\Spinner.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UI\OsuScoreOverlay.cs" />
     <Compile Include="OsuMod.cs" />
   </ItemGroup>
   <ItemGroup>
