+++ osu.Game.Modes.Taiko/TaikoRuleset.cs
@@ -1,20 +1,19 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
+using OpenTK.Input;
 using osu.Game.Beatmaps;
 using osu.Game.Graphics;
 using osu.Game.Modes.Objects;
-using osu.Game.Modes.Osu.UI;
 using osu.Game.Modes.Taiko.UI;
 using osu.Game.Modes.UI;
+using osu.Game.Screens.Play;
 using System.Collections.Generic;
 
 namespace osu.Game.Modes.Taiko
 {
     public class TaikoRuleset : Ruleset
     {
-        public override ScoreOverlay CreateScoreOverlay() => new OsuScoreOverlay();
-
         public override HitRenderer CreateHitRendererWith(Beatmap beatmap) => new TaikoHitRenderer
         {
             Beatmap = beatmap,
@@ -80,6 +79,14 @@ namespace osu.Game.Modes.Taiko
         protected override PlayMode PlayMode => PlayMode.Taiko;
 
         public override FontAwesome Icon => FontAwesome.fa_osu_taiko_o;
+
+        public override KeyCounter[] GameplayKeys => new KeyCounter[]
+        {
+            new KeyCounterKeyboard(Key.D),
+            new KeyCounterKeyboard(Key.F),
+            new KeyCounterKeyboard(Key.J),
+            new KeyCounterKeyboard(Key.K)
+        };
 
         public override ScoreProcessor CreateScoreProcessor(int hitObjectCount = 0) => null;
