+++ osu.Game/Modes/UI/StandardHUDOverlay.cs
@@ -1,0 +1,54 @@
+﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+
+using OpenTK;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Primitives;
+using osu.Game.Graphics.UserInterface;
+using osu.Game.Screens.Play;
+
+namespace osu.Game.Modes.UI
+{
+    internal class StandardHUDOverlay : HUDOverlay
+    {
+        public StandardHUDOverlay(Ruleset ruleset)
+            : base(ruleset)
+        {
+        }
+
+        protected override PercentageCounter CreateAccuracyCounter() => new PercentageCounter
+        {
+            Anchor = Anchor.TopCentre,
+            Origin = Anchor.TopCentre,
+            Position = new Vector2(0, 65),
+            TextSize = 20,
+            Margin = new MarginPadding { Right = 5 },
+        };
+
+        protected override BaseComboCounter CreateComboCounter() => new ComboCounter
+        {
+            Anchor = Anchor.BottomLeft,
+            Origin = Anchor.BottomLeft,
+        };
+
+        protected override KeyCounterCollection CreateKeyCounter(KeyCounter[] keyCounters) => new KeyCounterCollection
+        {
+            IsCounting = true,
+            FadeTime = 50,
+            Anchor = Anchor.BottomRight,
+            Origin = Anchor.BottomRight,
+            Margin = new MarginPadding(10),
+            Children = keyCounters
+        };
+
+        protected override ScoreCounter CreateScoreCounter() => new ScoreCounter(6)
+        {
+            Anchor = Anchor.TopCentre,
+            Origin = Anchor.TopCentre,
+            TextSize = 40,
+            Position = new Vector2(0, 30),
+            Margin = new MarginPadding { Right = 5 },
+        };
+    }
+}
